+++ src/App.tsx
@@ -3,6 +3,7 @@ import { useFormState } from "./hooks/useFormState";
 import { useAppState, type DeckData } from "./hooks/useAppState";
 import { useModalState } from "./hooks/useModalState";
 import { useLocalStorage } from "./hooks/useLocalStorage";
+import { useDeckManager } from "./hooks/useDeckManager";
 import { DeckCard } from "./components/DeckCard";
 import { FormInput } from "./components/FormInput";
 import { ImageModal } from "./components/ImageModal";
@@ -12,6 +13,13 @@ function App() {
   const [appState, appActions] = useAppState();
   const [modalState, modalActions] = useModalState(appState.deckList);
   const { loadDeckList, saveDeckList } = useLocalStorage();
+  const { handleSubmit, handleRemoveDeck, handleClearAll } = useDeckManager(
+    appState,
+    appActions,
+    formState,
+    formActions,
+    saveDeckList
+  );
 
   // 初期化時にlocalStorageからデータを読み込む
   useEffect(() => {
@@ -20,198 +28,6 @@ function App() {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
-  const addSingleDeck = (
-    code: string,
-    playerName?: string
-  ): DeckData | null => {
-    const trimmedCode = code.trim();
-    const trimmedPlayerName = playerName?.trim();
-
-    if (!trimmedCode) return null;
-
-    // 重複チェック
-    if (appState.deckList.some((deck) => deck.code === trimmedCode)) {
-      return null;
-    }
-
-    const imageUrl = `https://www.pokemon-card.com/deck/deckView.php/deckID/${trimmedCode}`;
-    return {
-      id: `${Date.now()}-${Math.random()}`,
-      code: trimmedCode,
-      playerName: trimmedPlayerName || undefined,
-      imageUrl,
-      addedAt: new Date(),
-    };
-  };
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-
-    if (formState.isBulkMode) {
-      if (!formState.bulkMode.input.trim()) {
-        appActions.setError("デッキコードを入力してください");
-        return;
-      }
-
-      await handleBulkSubmit();
-    } else {
-      if (!formState.singleMode.deckCode.trim()) {
-        appActions.setError("デッキコードを入力してください");
-        return;
-      }
-
-      await handleSingleSubmit();
-    }
-  };
-
-  const handleSingleSubmit = async () => {
-    const trimmedCode = formState.singleMode.deckCode.trim();
-
-    // 重複チェック
-    if (appState.deckList.some((deck) => deck.code === trimmedCode)) {
-      appActions.setError("このデッキコードは既に追加されています");
-      return;
-    }
-
-    appActions.setLoading(true);
-    appActions.setError(null);
-
-    try {
-      const newDeck = addSingleDeck(
-        trimmedCode,
-        formState.singleMode.playerName
-      );
-      if (newDeck) {
-        const updatedList = [newDeck, ...appState.deckList];
-        appActions.setDeckList(updatedList);
-        saveDeckList(updatedList);
-        formActions.resetSingleForm();
-      }
-    } catch {
-      appActions.setError("デッキレシピの取得に失敗しました");
-    } finally {
-      appActions.setLoading(false);
-    }
-  };
-
-  const handleBulkSubmit = async () => {
-    const lines = formState.bulkMode.input
-      .split("\n")
-      .map((line) => line.trim())
-      .filter((line) => line.length > 0);
-
-    if (lines.length === 0) {
-      appActions.setError("有効なデータが見つかりません");
-      return;
-    }
-
-    appActions.setLoading(true);
-    appActions.setError(null);
-    appActions.setProgress({ current: 0, total: lines.length });
-
-    const newDecks: DeckData[] = [];
-    const duplicates: string[] = [];
-    const errors: string[] = [];
-
-    try {
-      for (let i = 0; i < lines.length; i++) {
-        const line = lines[i];
-        appActions.setProgress({ current: i + 1, total: lines.length });
-
-        let playerName: string | undefined;
-        let code: string;
-
-        // タブ区切りかどうかチェック（Excelからのコピー）
-        if (line.includes("\t")) {
-          const parts = line.split("\t");
-          playerName = parts[0]?.trim();
-          code = parts[1]?.trim() || "";
-        } else {
-          // 他の区切り文字で分割を試す
-          const parts = line.split(/[,;\s]+/);
-          if (parts.length >= 2) {
-            playerName = parts[0]?.trim();
-            code = parts[1]?.trim() || "";
-          } else {
-            // 単一のデッキコードとして処理
-            code = parts[0]?.trim() || "";
-          }
-        }
-
-        if (!code) {
-          errors.push(`行 ${i + 1}: デッキコードが見つかりません`);
-          continue;
-        }
-
-        // 既存のリストまたは今回追加予定のリストに重複がないかチェック
-        const isDuplicate =
-          appState.deckList.some((deck) => deck.code === code) ||
-          newDecks.some((deck) => deck.code === code);
-
-        if (isDuplicate) {
-          duplicates.push(code);
-          continue;
-        }
-
-        const newDeck = addSingleDeck(code, playerName);
-        if (newDeck) {
-          newDecks.push(newDeck);
-        }
-
-        // 処理間隔を設ける（UI更新のため）
-        if (i < lines.length - 1) {
-          await new Promise((resolve) => setTimeout(resolve, 100));
-        }
-      }
-
-      if (newDecks.length > 0) {
-        const updatedList = [...newDecks, ...appState.deckList];
-        appActions.setDeckList(updatedList);
-        saveDeckList(updatedList);
-      }
-
-      let message = "";
-      if (newDecks.length > 0) {
-        message += `${newDecks.length}件を追加しました。`;
-      }
-      if (duplicates.length > 0) {
-        message += ` ${duplicates.length}件のデッキコードは既に追加済みのためスキップしました。`;
-      }
-      if (errors.length > 0) {
-        message += ` ${errors.length}件でエラーが発生しました。`;
-      }
-
-      if (
-        newDecks.length === 0 &&
-        (duplicates.length > 0 || errors.length > 0)
-      ) {
-        appActions.setError(
-          message || "追加できるデッキが見つかりませんでした"
-        );
-      } else if (duplicates.length > 0 || errors.length > 0) {
-        appActions.setError(message);
-      } else {
-        formActions.resetBulkForm();
-      }
-    } catch {
-      appActions.setError("一括処理中にエラーが発生しました");
-    } finally {
-      appActions.setLoading(false);
-      appActions.setProgress(null);
-    }
-  };
-
-  const handleRemoveDeck = (id: string) => {
-    appActions.removeDeck(id);
-    saveDeckList(appState.deckList.filter((deck) => deck.id !== id));
-  };
-
-  const handleClearAll = () => {
-    appActions.clearAll();
-    saveDeckList([]);
-    formActions.resetForm();
-    appActions.setError(null);
-  };
 
   const handleSingleModeSwitch = () => {
     formActions.setSingleMode();
